refactor: logging, addon factory, batched ASG, scale-dry-run PDBs, split actions.go (REF-52/8 follow-ons)
Batch F of the Sonnet 4.6 backlog — Phase 8 refactors:

- REF-37: single logger path. factory.NewDefaultLogger now reads a configurable
  level (factory.SetDefaultLogLevel/ParseLogLevel); global --log-level
  (debug|info|warn|error, default warn) and --verbose flags wired in main's
  Before. Removed the ad-hoc slog.New(...LevelError) in nodegroup list.
- REF-39: factory.NewAddonService; addon actions build the service through the
  factory (and resolve names via addonSvc.EKS()) — no direct eks.NewFromConfig
  in addon/actions.go.
- REF-50: health capacity check collects all ASG names across nodegroups and
  resolves instances via chunked (≤100) DescribeAutoScalingGroups instead of one
  describe per ASG.
- REF-4: `nodegroup scale --dry-run --check-pdbs` now lists the specific PDBs
  that would constrain a scale-down (namespace/name, disruptionsAllowed,
  healthy counts) via a new health.ListPodDisruptionBudgets, or confirms none do.
- REF-38: split the 756-line nodegroup/actions.go by concern into
  actions_list.go, actions_describe.go, actions_scale.go, actions_update.go
  (shared resolveHealthKubeClient stays in actions.go). No behavior change.

Tests: ParseLogLevel/SetDefaultLogLevel, NewAddonService. Full gate green
(build, gofmt, go vet, go test -race ./..., golangci-lint 0).
