+++ internal/commands/factory/factory.go
@@ -4,6 +4,7 @@ package factory
 import (
 	"log/slog"
 	"os"
+	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/autoscaling"
@@ -12,17 +13,45 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/dantech2000/refresh/internal/health"
+	"github.com/dantech2000/refresh/internal/services/addons"
 	"github.com/dantech2000/refresh/internal/services/cluster"
 	"github.com/dantech2000/refresh/internal/services/nodegroup"
 )
 
+// defaultLogLevel is the level used by NewDefaultLogger when a command does not
+// supply its own logger. It defaults to warn (quiet) and is overridden once,
+// from main's Before hook, by the global --log-level / --verbose flags (REF-37).
+var defaultLogLevel = slog.LevelWarn
+
+// SetDefaultLogLevel sets the level NewDefaultLogger uses for the shared logger.
+// Call it once during startup; commands run single-threaded after that.
+func SetDefaultLogLevel(level slog.Level) { defaultLogLevel = level }
+
+// ParseLogLevel maps a --log-level string to an slog.Level. Unknown values
+// fall back to warn (the default quiet level). (REF-37)
+func ParseLogLevel(s string) slog.Level {
+	switch strings.ToLower(strings.TrimSpace(s)) {
+	case "debug":
+		return slog.LevelDebug
+	case "info":
+		return slog.LevelInfo
+	case "warn", "warning", "":
+		return slog.LevelWarn
+	case "error":
+		return slog.LevelError
+	default:
+		return slog.LevelWarn
+	}
+}
+
 // NewDefaultLogger returns logger unchanged if non-nil; otherwise returns a
-// stderr warn-level text logger.
+// stderr text logger at the configured default level (see SetDefaultLogLevel).
+// This is the single logger-construction path for the CLI.
 func NewDefaultLogger(logger *slog.Logger) *slog.Logger {
 	if logger != nil {
 		return logger
 	}
-	return slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: slog.LevelWarn}))
+	return slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: defaultLogLevel}))
 }
 
 // NewClusterService initializes a cluster service with optional health checking.
@@ -49,6 +78,12 @@ func NewNodegroupService(awsCfg aws.Config, withHealth bool, logger *slog.Logger
 		hc = health.NewChecker(eksClient, nil, cwClient, asgClient)
 	}
 	return nodegroup.NewService(awsCfg, hc, logger)
+}
+
+// NewAddonService initializes an add-on service through the shared logger path,
+// matching the cluster/nodegroup constructors. (REF-39)
+func NewAddonService(awsCfg aws.Config, logger *slog.Logger) *addons.ServiceImpl {
+	return addons.NewService(eks.NewFromConfig(awsCfg), NewDefaultLogger(logger))
 }
 
 // NewNodegroupServiceWithHealth initializes a nodegroup service whose health
